requirements.txt
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.1.1
flake8==5.0.4
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
greenlet==1.1.3.post0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.16.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.3
jupyter_core==4.11.2
mando==0.6.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mkdocs-material-extensions==1.1
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.10.0
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==4.21.11
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pyzmq==24.0.1
radon==5.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.42
stack-data==0.5.1
tenacity==8.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
